@@ -40,8 +40,8 @@
         "enlightn/security-checker": "^2.0",
         "guzzlehttp/guzzle": "^7.0",
         "larastan/larastan": "^3.0",
-        "laravel/framework": "^9.0|^10.0|^11.0|^12.0",
-        "nikic/php-parser": "^4.0|^5.0",
+        "laravel/framework": "^12.40",
+        "nikic/php-parser": "^5.0",
         "phpstan/phpstan": "^2.0",
         "phpstan/phpstan-deprecation-rules": "^2.0",
         "symfony/finder": "^4.0|^5.0|^6.0|^7.0"
@@ -51,8 +51,8 @@
         "brianium/paratest": "^6.1|^7.0",
         "friendsofphp/php-cs-fixer": "^2.18|^3.0",
         "mockery/mockery": "^1.3",
-        "orchestra/testbench": "^7.0|^8.0|^9.0",
-        "phpunit/phpunit": "^9.0|^10.0",
+        "orchestra/testbench": "^10.8",
+        "phpunit/phpunit": "^12",
         "predis/predis": "*"
     },
     "autoload": {
