@@ -4,19 +4,18 @@ namespace Enlightn\Enlightn\Tests\Analyzers\Performance;
 
 use Enlightn\Enlightn\Analyzers\Performance\AutoloaderOptimizationAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 class AutoloaderOptimizationAnalyzerTest extends AnalyzerTestCase
 {
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(AutoloaderOptimizationAnalyzer::class, $app);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function skips_in_local()
     {
         $this->app->config->set('app.env', 'local');
@@ -26,9 +25,7 @@ class AutoloaderOptimizationAnalyzerTest extends AnalyzerTestCase
         $this->assertSkipped(AutoloaderOptimizationAnalyzer::class);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_non_optimized_autoloader()
     {
         $this->app->config->set('app.env', 'production');
