@@ -4,35 +4,31 @@ namespace Enlightn\Enlightn\Tests\Analyzers\Performance;
 
 use Enlightn\Enlightn\Analyzers\Performance\ConfigCachingAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 class ConfigCachingAnalyzerTest extends AnalyzerTestCase
 {
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(ConfigCachingAnalyzer::class, $app);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_cached_config_in_local()
     {
         $this->app->config->set('app.env', 'local');
 
-        touch($this->app->getCachedConfigPath());
+        // Set the config_loaded_from_cache binding to simulate cached config
+        $this->app->instance('config_loaded_from_cache', true);
 
         $this->runEnlightn();
 
         $this->assertFailed(ConfigCachingAnalyzer::class);
-
-        unlink($this->app->getCachedConfigPath());
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_non_cached_config_in_production()
     {
         $this->app->config->set('app.env', 'production');
@@ -42,25 +38,20 @@ class ConfigCachingAnalyzerTest extends AnalyzerTestCase
         $this->assertFailed(ConfigCachingAnalyzer::class);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_cached_config_in_production()
     {
         $this->app->config->set('app.env', 'production');
 
-        touch($this->app->getCachedConfigPath());
+        // Set the config_loaded_from_cache binding to simulate cached config
+        $this->app->instance('config_loaded_from_cache', true);
 
         $this->runEnlightn();
 
         $this->assertPassed(ConfigCachingAnalyzer::class);
-
-        unlink($this->app->getCachedConfigPath());
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_non_cached_config_in_local()
     {
         $this->app->config->set('app.env', 'local');
