@@ -4,19 +4,18 @@ namespace Enlightn\Enlightn\Tests\Analyzers\Performance;
 
 use Enlightn\Enlightn\Analyzers\Performance\QueueDriverAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 class QueueDriverAnalyzerTest extends AnalyzerTestCase
 {
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(QueueDriverAnalyzer::class, $app);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_null_queue_driver()
     {
         $this->app->config->set('queue.default', 'null');
@@ -26,9 +25,7 @@ class QueueDriverAnalyzerTest extends AnalyzerTestCase
         $this->assertFailedAt(QueueDriverAnalyzer::class, $this->getConfigStubPath('queue'), 16);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_sync_queue_driver()
     {
         $this->app->config->set('queue.default', 'sync');
@@ -38,9 +35,7 @@ class QueueDriverAnalyzerTest extends AnalyzerTestCase
         $this->assertFailedAt(QueueDriverAnalyzer::class, $this->getConfigStubPath('queue'), 16);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_database_queue_driver_in_non_local_env()
     {
         $this->app->config->set('queue.default', 'database');
@@ -51,9 +46,7 @@ class QueueDriverAnalyzerTest extends AnalyzerTestCase
         $this->assertFailedAt(QueueDriverAnalyzer::class, $this->getConfigStubPath('queue'), 16);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_database_queue_driver_in_local_env()
     {
         $this->app->config->set('queue.default', 'database');
