@@ -4,35 +4,31 @@ namespace Enlightn\Enlightn\Tests\Analyzers\Performance;
 
 use Enlightn\Enlightn\Analyzers\Performance\RouteCachingAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 class RouteCachingAnalyzerTest extends AnalyzerTestCase
 {
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(RouteCachingAnalyzer::class, $app);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_cached_routes_in_local()
     {
         $this->app->config->set('app.env', 'local');
         
-        touch($this->app->getCachedRoutesPath());
+        // Set the routes.cached binding to simulate cached routes
+        $this->app->instance('routes.cached', true);
 
         $this->runEnlightn();
 
         $this->assertFailed(RouteCachingAnalyzer::class);
-        
-        unlink($this->app->getCachedRoutesPath());
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_non_cached_routes_in_production()
     {
         $this->app->config->set('app.env', 'production');
@@ -42,25 +38,20 @@ class RouteCachingAnalyzerTest extends AnalyzerTestCase
         $this->assertFailed(RouteCachingAnalyzer::class);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_cached_routes_in_production()
     {
         $this->app->config->set('app.env', 'production');
         
-        touch($this->app->getCachedRoutesPath());
+        // Set the routes.cached binding to simulate cached routes
+        $this->app->instance('routes.cached', true);
 
         $this->runEnlightn();
 
         $this->assertPassed(RouteCachingAnalyzer::class);
-
-        unlink($this->app->getCachedRoutesPath());
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_non_cached_routes_in_local()
     {
         $this->app->config->set('app.env', 'local');
