@@ -6,14 +6,15 @@ use Enlightn\Enlightn\Analyzers\Reliability\EnvFileAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
 use Illuminate\Filesystem\Filesystem;
 use Mockery as m;
+use PHPUnit\Framework\Attributes\Test;
 
 class EnvFileAnalyzerTest extends AnalyzerTestCase
 {
     protected $files;
 
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(EnvFileAnalyzer::class, $app);
 
@@ -24,9 +25,7 @@ class EnvFileAnalyzerTest extends AnalyzerTestCase
         });
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_with_env_file()
     {
         $this->files->shouldReceive('exists')->with(base_path('.env'))->andReturn(true);
@@ -36,9 +35,7 @@ class EnvFileAnalyzerTest extends AnalyzerTestCase
         $this->assertPassed(EnvFileAnalyzer::class);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_missing_env_file()
     {
         $this->files->shouldReceive('exists')->with(base_path('.env'))->andReturn(false);
