@@ -5,21 +5,20 @@ namespace Enlightn\Enlightn\Tests\Analyzers\Reliability;
 use Enlightn\Enlightn\Analyzers\Reliability\EnvVariableAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
 use Enlightn\Enlightn\Tests\Stubs\EnvStub;
+use PHPUnit\Framework\Attributes\Test;
 
 class EnvVariableAnalyzerTest extends AnalyzerTestCase
 {
     protected $files;
 
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(EnvVariableAnalyzer::class, $app);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_missing_env_variables()
     {
         $this->app->setBasePath(dirname($this->getClassStubPath(EnvStub::class)));
