@@ -6,19 +6,18 @@ use Enlightn\Enlightn\Analyzers\Reliability\ForeachIterableAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
 use Enlightn\Enlightn\Tests\Stubs\DummyStub;
 use Enlightn\Enlightn\Tests\Stubs\ForeachIterableStub;
+use PHPUnit\Framework\Attributes\Test;
 
 class ForeachIterableAnalyzerTest extends AnalyzerTestCase
 {
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(ForeachIterableAnalyzer::class, $app);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_non_iterable_foreach()
     {
         $this->setBasePathFrom(ForeachIterableStub::class);
@@ -30,9 +29,7 @@ class ForeachIterableAnalyzerTest extends AnalyzerTestCase
         $this->assertHasErrors(ForeachIterableAnalyzer::class, 2);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_with_no_foreach()
     {
         $this->setBasePathFrom(DummyStub::class);
