@@ -6,19 +6,18 @@ use Enlightn\Enlightn\Analyzers\Reliability\InvalidOffsetAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
 use Enlightn\Enlightn\Tests\Stubs\DummyStub;
 use Enlightn\Enlightn\Tests\Stubs\InvalidOffsetStub;
+use PHPUnit\Framework\Attributes\Test;
 
 class InvalidOffsetAnalyzerTest extends AnalyzerTestCase
 {
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(InvalidOffsetAnalyzer::class, $app);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_invalid_offset()
     {
         $this->setBasePathFrom(InvalidOffsetStub::class);
@@ -32,9 +31,7 @@ class InvalidOffsetAnalyzerTest extends AnalyzerTestCase
         $this->assertHasErrors(InvalidOffsetAnalyzer::class, 5);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_with_no_offset()
     {
         $this->setBasePathFrom(DummyStub::class);
