@@ -6,19 +6,18 @@ use Enlightn\Enlightn\Analyzers\Reliability\MissingModelRelationAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
 use Enlightn\Enlightn\Tests\Stubs\DummyStub;
 use Enlightn\Enlightn\Tests\Stubs\MissingRelationStub;
+use PHPUnit\Framework\Attributes\Test;
 
 class MissingModelRelationAnalyzerTest extends AnalyzerTestCase
 {
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(MissingModelRelationAnalyzer::class, $app);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_missing_relations()
     {
         $this->setBasePathFrom(MissingRelationStub::class);
@@ -30,9 +29,7 @@ class MissingModelRelationAnalyzerTest extends AnalyzerTestCase
         $this->assertHasErrors(MissingModelRelationAnalyzer::class, 2);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_with_no_relations()
     {
         $this->setBasePathFrom(DummyStub::class);
