@@ -5,19 +5,18 @@ namespace Enlightn\Enlightn\Tests\Analyzers\Reliability;
 use Enlightn\Enlightn\Analyzers\Reliability\SyntaxErrorAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
 use Enlightn\Enlightn\Tests\Stubs\DummyStub;
+use PHPUnit\Framework\Attributes\Test;
 
 class SyntaxErrorAnalyzerTest extends AnalyzerTestCase
 {
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(SyntaxErrorAnalyzer::class, $app);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_syntax_errors()
     {
         $this->app->config->set('enlightn.base_path', $this->getBaseStubPath());
@@ -31,9 +30,7 @@ class SyntaxErrorAnalyzerTest extends AnalyzerTestCase
         $this->assertHasErrors(SyntaxErrorAnalyzer::class, 2);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_with_no_errors()
     {
         $this->setBasePathFrom(DummyStub::class);
