@@ -6,19 +6,18 @@ use Enlightn\Enlightn\Analyzers\Reliability\UndefinedConstantAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
 use Enlightn\Enlightn\Tests\Stubs\DummyStub;
 use Enlightn\Enlightn\Tests\Stubs\UndefinedConstantStub;
+use PHPUnit\Framework\Attributes\Test;
 
 class UndefinedConstantAnalyzerTest extends AnalyzerTestCase
 {
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(UndefinedConstantAnalyzer::class, $app);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_missing_return_statements()
     {
         $this->setBasePathFrom(UndefinedConstantStub::class);
@@ -29,9 +28,7 @@ class UndefinedConstantAnalyzerTest extends AnalyzerTestCase
         $this->assertHasErrors(UndefinedConstantAnalyzer::class, 1);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_with_no_return_statements()
     {
         $this->setBasePathFrom(DummyStub::class);
