@@ -6,12 +6,13 @@ use Enlightn\Enlightn\Analyzers\Reliability\UpToDateMigrationsAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
 use Enlightn\Enlightn\Tests\Kernel;
 use Illuminate\Support\Facades\Artisan;
+use PHPUnit\Framework\Attributes\Test;
 
 class UpToDateMigrationsAnalyzerTest extends AnalyzerTestCase
 {
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(UpToDateMigrationsAnalyzer::class, $app);
     }
@@ -27,9 +28,7 @@ class UpToDateMigrationsAnalyzerTest extends AnalyzerTestCase
         $app->singleton('Illuminate\Contracts\Console\Kernel', Kernel::class);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_with_no_remaining_migrations()
     {
         Artisan::shouldReceive('call');
@@ -40,9 +39,7 @@ class UpToDateMigrationsAnalyzerTest extends AnalyzerTestCase
         $this->assertPassed(UpToDateMigrationsAnalyzer::class);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_pending_migrations()
     {
         Artisan::shouldReceive('call');
