@@ -4,19 +4,18 @@ namespace Enlightn\Enlightn\Tests\Analyzers\Security;
 
 use Enlightn\Enlightn\Analyzers\Security\AppDebugAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 class AppDebugAnalyzerTest extends AnalyzerTestCase
 {
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(AppDebugAnalyzer::class, $app);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_app_debug_in_production()
     {
         $this->app->config->set('app.env', 'production');
@@ -27,9 +26,7 @@ class AppDebugAnalyzerTest extends AnalyzerTestCase
         $this->assertFailedAt(AppDebugAnalyzer::class, $this->getConfigStubPath('app'), 42);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_with_app_debug_in_local()
     {
         $this->app->config->set('app.env', 'local');
@@ -40,9 +37,7 @@ class AppDebugAnalyzerTest extends AnalyzerTestCase
         $this->assertPassed(AppDebugAnalyzer::class);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_without_app_debug()
     {
         $this->app->config->set('app.env', 'production');
