@@ -6,19 +6,18 @@ use Enlightn\Enlightn\Analyzers\Security\FillableForeignKeyAnalyzer;
 use Enlightn\Enlightn\Tests\Analyzers\AnalyzerTestCase;
 use Enlightn\Enlightn\Tests\Stubs\DummyStub;
 use Enlightn\Enlightn\Tests\Stubs\FillableForeignKeyStub;
+use PHPUnit\Framework\Attributes\Test;
 
 class FillableForeignKeyAnalyzerTest extends AnalyzerTestCase
 {
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
-        parent::getEnvironmentSetUp($app);
+        parent::defineEnvironment($app);
 
         $this->setupEnvironmentFor(FillableForeignKeyAnalyzer::class, $app);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function detects_validation_sql_injection()
     {
         $this->setBasePathFrom(FillableForeignKeyStub::class);
@@ -29,9 +28,7 @@ class FillableForeignKeyAnalyzerTest extends AnalyzerTestCase
         $this->assertHasErrors(FillableForeignKeyAnalyzer::class, 1);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function passes_with_no_injection_call()
     {
         $this->setBasePathFrom(DummyStub::class);
