@@ -75,8 +75,8 @@ class QueueTimeoutAnalyzer extends ReliabilityAnalyzer
     {
         $connections = collect($config->get('queue.connections', []))
             ->filter(function ($conf, $queue) {
-                // skip sqs and sync drivers as they don't have retry after values
-                return ! in_array($conf['driver'], ['sqs', 'sync']);
+                // skip sqs, sync, deferred and failover drivers as they don't have retry after values
+                return ! in_array($conf['driver'], ['sqs', 'sync', 'deferred', 'failover']);
             })->map(function ($conf, $queue) {
                 return $this->getTimeoutAndRetryAfter($conf);
             })->filter(function ($conf) {
