@@ -2,7 +2,7 @@
 
 namespace Enlightn\Enlightn\Inspection\Concerns;
 
-use PhpParser\Node\Stmt\UseUse;
+use PhpParser\Node\UseItem;
 
 trait GathersClassAliases
 {
@@ -14,7 +14,7 @@ trait GathersClassAliases
     public function beforeTraverse(array $nodes)
     {
         foreach ($nodes as $node) {
-            if ($node instanceof UseUse
+            if ($node instanceof UseItem
                 && ($node->name->toString() === $this->class
                     || is_subclass_of($node->name->toString(), $this->class))) {
                 // Register the use statement and add the alias. E.g. "use ClassName as Alias" or "use ClassName".
