@@ -7,21 +7,22 @@ namespace Enlightn\Enlightn\PHPStan;
 use PHPStan\TrinaryLogic;
 use PHPStan\Type\ArrayType;
 use PHPStan\Type\CompoundType;
+use PHPStan\Type\IsSuperTypeOfResult;
 use PHPStan\Type\Type;
 use PHPStan\Type\UnionType;
 
 class RequestArrayDataType extends ArrayType
 {
-    public function isSuperTypeOf(Type $type): TrinaryLogic
+    public function isSuperTypeOf(Type $type): IsSuperTypeOfResult
     {
         if ($type instanceof self) {
-            return TrinaryLogic::createYes();
+            return IsSuperTypeOfResult::createYes();
         }
         if ($type instanceof CompoundType) {
             return $type->isSubTypeOf($this);
         }
 
-        return TrinaryLogic::createNo();
+        return IsSuperTypeOfResult::createNo();
     }
 
     public function canBeSuperTypeOf(Type $type): TrinaryLogic
