@@ -6,6 +6,7 @@ namespace Enlightn\Enlightn\PHPStan;
 
 use PHPStan\TrinaryLogic;
 use PHPStan\Type\CompoundType;
+use PHPStan\Type\IsSuperTypeOfResult;
 use PHPStan\Type\StringType;
 use PHPStan\Type\Type;
 use PHPStan\Type\UnionType;
@@ -21,16 +22,16 @@ class RequestDataType extends StringType
         return new self();
     }
 
-    public function isSuperTypeOf(Type $type) : TrinaryLogic
+    public function isSuperTypeOf(Type $type): IsSuperTypeOfResult
     {
         if ($type instanceof self) {
-            return TrinaryLogic::createYes();
+            return IsSuperTypeOfResult::createYes();
         }
         if ($type instanceof CompoundType) {
             return $type->isSubTypeOf($this);
         }
 
-        return TrinaryLogic::createNo();
+        return IsSuperTypeOfResult::createNo();
     }
 
     public function canBeSuperTypeOf(Type $type): TrinaryLogic
